requirements.txt
arrow==1.2.3
asgiref==3.6.0
autopep8==2.0.1
bcrypt==4.0.1
bfg==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.6
greenlet==2.0.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
numpy==1.24.2
pandas==1.5.3
passlib==1.7.4
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.2
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
tomli==2.0.1
twilio==7.16.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
